LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-es-expected.txt

PASS navigator.language is "es"
PASS successfullyParsed is true

TEST COMPLETE
